rename helpers folder for app-queries
